+++ projects/kypo-sandbox-agenda/sandbox-resources/src/models/vm-images-row-adapter.ts
@@ -3,4 +3,6 @@ import { VirtualImage } from '@muni-kypo-crp/sandbox-model';
 export class VMImagesRowAdapter extends VirtualImage {
   updatedAtFormatted: string;
   createdAtFormatted: string;
+  guiAccessFormatted: string;
+  sizeFormatted: number;
 }