+++ projects/kypo-sandbox-agenda/sandbox-resources/src/services/vm-images-concrete.service.ts
@@ -7,7 +7,6 @@ import { Injectable } from '@angular/core';
 import { VirtualImage } from '@muni-kypo-crp/sandbox-model';
 import { PaginatedResource, OffsetPaginationEvent, SentinelFilter } from '@sentinel/common';
 import { VMImagesService } from './vm-images.service';
-import { HttpParams } from '@angular/common/http';
 
 @Injectable()
 export class VMImagesConcreteService extends VMImagesService {
@@ -22,16 +21,22 @@ export class VMImagesConcreteService extends VMImagesService {
   /**
    * Retrieves paginated available virtual machine images
    * @param pagination requested pagination
+   * @param onlyKypoImages filters images belonging to KYPO
+   * @param onlyGuiAccess filters images with GUI access
+   * @param cached Performs the faster version of this endpoint but does not retrieve a fresh list of images
+   * @param filter list of sentinel filters to filter results
+   *
    */
   getAvailableImages(
     pagination: OffsetPaginationEvent,
     onlyKypoImages?: boolean,
+    onlyGuiAccess?: boolean,
     cached?: boolean,
     filter?: string
   ): Observable<PaginatedResource<VirtualImage>> {
     this.isLoadingSubject$.next(true);
     const filters = filter ? [new SentinelFilter('name', filter)] : [];
-    return this.vmImagesApi.getAvailableImages(pagination, onlyKypoImages, cached, filters).pipe(
+    return this.vmImagesApi.getAvailableImages(pagination, onlyKypoImages, onlyGuiAccess, cached, filters).pipe(
       tap(
         (resource) => {
           this.resourceSubject$.next(resource);